Projects New Relic supports via GitHub Sponsors  - Homepage
On the homepage I added an additional section below "Explore our projects". It is styled the same as the "Explore our projects" panel, with 8 randomly selected folks shown on the homepage and a "view all" link which takes you to https://github.com/orgs/newrelic/sponsoring.

The card for each person is their avatar, name, and bio from GitHub.