Fix CI badge in README
The CI badge in the README was showing a failed indicator. I believe this is because there is no workflow named CI.yml, rather the project uses a `pr.yml` file for a CI check. When we update the README to point to the correct workflow, the badge accurately reflects its status.